requirements.txt
langchain==0.2.1
langchain_community==0.2.1
langchain_openai==0.1.7
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.32.2
faiss-cpu
faiss-cpu==1.8.0
pillow==10.2.0
